refactor(client-3d): extract shared useWallOcclusion hook
Phase 3E: Consolidated near-identical wall-fade logic from Room.tsx,
JukeboxRoom.tsx, and JapaneseRoom.tsx into a single useWallOcclusion
hook. Each room dropped ~100 LOC of duplicated raycaster + material
mutation code (Room 411→313, JukeboxRoom 417→320, JapaneseRoom
417→331). The hook handles ShaderMaterial via uOpacity, emissive
scaling for neon/glow props, and the DoubleSide raycast hack for
backface hits.
